pip_requirements.txt
Babel==2.7.0
coverage==4.5.3
Django==2.2.3
django-phonenumber-field==3.0.1
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
phonenumbers==8.10.14
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
